Re-load the native library if it was already loaded.
Java doesn't like loading the same native library twice from the same
location in different classloaders. We force it by copying the library
and loading that.

A better solution would be to load the library in the top-level
classloader, but we may not have access to it, since System.loadLibrary
is caller-sensitive, meaning it loads the library into the classloader of
the calling class.